Add: View de Veiculo e Categoria

+++ Locadora.Models/Categoria.cs
@@ -30,16 +30,16 @@ namespace Locadora.Models
         public int CategoriaId { get; private set; }
         public string Nome { get; private set; }
         public string? Descricao { get; private set; }
-        public double Diaria { get; private set; }
+        public decimal Diaria { get; private set; }
 
-        public Categoria(string nome, string? descricao, double diaria)
+        public Categoria(string nome, string? descricao, decimal diaria)
         {
             Nome = nome;
             Descricao = descricao;
             Diaria = diaria;
         }
 
-        public void setDiaria(double diaria)
+        public void setDiaria(decimal diaria)
         {
             Diaria = diaria;
         }
